:hammer: Change the way that replies are detected and handled
It seems that you can have a Tweet that is a reply to a user, but isn't a
reply to a Tweet. This seems to be the case where you've tweeted @ someone
but not as an actual reply; so the account is seen as an account you're
replying to, but there is no tweet you're replying to.